Autolog filename can be dir basename
If automatic logfiles are requested via env var

DCP_INSPECT_AUTOLOG

the generated autolog filenames will be a composite
of path elements, time info, app version and a short
random string - with the purpose of creating a unique
autolog file for every inspection run.

If, in addition, an env var called

DCP_INSPECT_AUTOLOG_NAME_IS_BASENAME

is present, then the filename for the generated autolog
will be the argument provided for inspection.

Suggesting to use this only if the user wants exactly
that behaviour and is aware of the implications.